[RB] - consolidate the feature switch functionality for live chat into a url param / session storage solution
